requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
Authlib==1.3.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
cycler==0.12.1
Django==4.2.9
django-admin-datta==1.0.10
django-api-generator==1.0.17
django-debug-toolbar==4.2.0
django-dynamic-datatb==1.0.25
djangorestframework==3.14.0
dparse==0.6.4b0
fonttools==4.50.0
frozenlist==1.4.1
fuzzywuzzy==0.18.0
gunicorn==21.2.0
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib
mdurl==0.1.2
multidict==6.0.5
numpy
packaging==24.0
pandas
pillow==10.2.0
pycparser==2.21
pydantic==1.10.14
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nmap==0.7.1
pytz==2024.1
requests==2.31.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==3.0.1
safety-schemas==0.0.2
six==1.16.0
sqlmap==1.8.3
sqlparse==0.4.4
str2bool==1.1
tld==0.13
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vt-py==0.18.0
vulnx==1.6
whitenoise==6.6.0
xsstrike==3.2.2
yarl==1.9.4
